requirements.txt
Flask==3.1.0
gunicorn==23.0.0
python-dotenv==1.1.0
google-api-core==2.24.2
google-auth==2.38.0
google-cloud-core==2.4.3
google-cloud-datastore==2.20.2
googleapis-common-protos==1.69.2
